Fixed loading of subscription data with the iron-router routes.
Now subscription data is loaded in the client before the template gets rendered
so that data is displayed correctly.  Before you would often see no data
available messages when displaying graphs on the home page and lists on the
individual and team stat pages would not sort properly.  I also added
subs-manager to cache subscriptions so that the data does not reload when
routes change.